lightly re-organized server
