+++ package.json
@@ -12,7 +12,6 @@
     "cors": "^2.8.5",
     "express": "^4.18.2",
     "fake-useragent": "^1.0.1",
-    "http-proxy-middleware": "^2.0.6",
     "https": "^1.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
